// src/models/recipe.ts
import { IFeedModel } from "./feeds";

export interface IRecipeModel {
    id: number;
    name: string;
    ingredients: string[];
    instructions: string[];
    prepTimeMinutes: number;
    cookTimeMinutes: number;
    servings: number;
    difficulty: string;
    cuisine: string;
    caloriesPerServing: number;
    tags: string[];
    userId: number;
    image: string;
    rating: number;
    reviewCount: number;
    mealType: string[];
}

export interface IRecipesFeedModel extends IFeedModel {
    recipes: IRecipeModel[];
}